This squash improves the accuracy of inference on the GoogLeNet model on ImageNet data.

FP32: avg top1 accuracy: 0.7050

Using INT8 and mkldnn accuracy increases from
INT8: avg top1 accuracy: 0.7017
to
INT8: avg top1 accuracy: 0.7022

This is reopen of  https://github.com/PaddlePaddle/Paddle/pull/18676

test=develop