@luotao1 2019-07-30
before:
```
I0730 18:03:55.057801 45052 tester_helper.h:462] --- Performance summary ---
I0730 18:03:55.057843 45052 tester_helper.h:463] FP32: avg fps: 33.5400, avg latency: 29.8152 ms
I0730 18:03:55.057854 45052 tester_helper.h:466] INT8: avg fps: 68.2523, avg latency: 14.6515 ms
I0730 18:03:55.067595 45052 tester_helper.h:447] --- Accuracy summary ---
I0730 18:03:55.067620 45052 tester_helper.h:448] Accepted top1 accuracy drop threshold: 0.01. (condition: (FP32_top1_acc - INT8_top1_acc) <= threshold)
I0730 18:03:55.067639 45052 tester_helper.h:451] FP32: avg top1 accuracy: 0.7050
I0730 18:03:55.067646 45052 tester_helper.h:453] INT8: avg top1 accuracy: 0.7008
```
after:
```
I0730 20:00:23.235113 294204 tester_helper.h:462] --- Performance summary ---
I0730 20:00:23.235152 294204 tester_helper.h:463] FP32: avg fps: 34.1651, avg latency: 29.2696 ms
I0730 20:00:23.235173 294204 tester_helper.h:466] INT8: avg fps: 82.6219, avg latency: 12.1033 ms
I0730 20:00:23.244776 294204 tester_helper.h:447] --- Accuracy summary ---
I0730 20:00:23.244801 294204 tester_helper.h:448] Accepted top1 accuracy drop threshold: 0.01. (condition: (FP32_top1_acc - INT8_top1_acc) <= threshold)
I0730 20:00:23.244819 294204 tester_helper.h:451] FP32: avg top1 accuracy: 0.7050
I0730 20:00:23.244824 294204 tester_helper.h:453] INT8: avg top1 accuracy: 0.7003
```
from 70.08->70.03