@bingyanghuang 2019-07-31
Based on LuoTao's benchmark, performance increase from 68.2523 to 82.6219, and the accuracy only drop from 0.7008 to 0.7003. I think this PR is good to merge. What do you think about? @wojtuss