@wozna 2019-07-31
I checked it on `Intel(R) Xeon(R) Gold 6248 CPU @ 2.50GHz` 
and I got the same results. 

before
`I0731 11:36:49.323259 42280 tester_helper.h:462] --- Performance summary ---
I0731 11:36:49.323282 42280 tester_helper.h:463] FP32: avg fps: 582.5014, avg latency: 1.7167 ms
I0731 11:36:49.323288 42280 tester_helper.h:466] INT8: avg fps: 1010.2657, avg latency: 0.9898 ms
I0731 11:36:49.323566 42280 tester_helper.h:447] --- Accuracy summary ---
I0731 11:36:49.323572 42280 tester_helper.h:448] Accepted top1 accuracy drop threshold: 0.01. (condition: (FP32_top1_acc - INT8_top1_acc) <= threshold)
I0731 11:36:49.323578 42280 tester_helper.h:451] FP32: avg top1 accuracy: 0.7050
I0731 11:36:49.323582 42280 tester_helper.h:453] INT8: avg top1 accuracy: 0.7017`
after
`I0731 11:27:40.606056 39099 tester_helper.h:462] --- Performance summary ---
I0731 11:27:40.606074 39099 tester_helper.h:463] FP32: avg fps: 600.3001, avg latency: 1.6658 ms
I0731 11:27:40.606081 39099 tester_helper.h:466] INT8: avg fps: 1146.9660, avg latency: 0.8719 ms
I0731 11:27:40.606254 39099 tester_helper.h:447] --- Accuracy summary ---
I0731 11:27:40.606259 39099 tester_helper.h:448] Accepted top1 accuracy drop threshold: 0.01. (condition: (FP32_top1_acc - INT8_top1_acc) <= threshold)
I0731 11:27:40.606263 39099 tester_helper.h:451] FP32: avg top1 accuracy: 0.7050
I0731 11:27:40.606267 39099 tester_helper.h:453] INT8: avg top1 accuracy: 0.7022`