@luotao1 2019-07-31
Could you provide:
- commit_id
- cmake command

Besides, I run with `OMP_NUM_THREADS=1`.