@wozna 2019-07-31
> Could you provide:
> 
> * commit_id 
16a09d5

> * cmake command
`/Paddle/build/paddle/fluid/inference/tests/api/test_analyzer_int8_image_classification "ARGS" 
"--infer_model=/Paddle/build/third_party/inference_demo/int8v2/googlenet/model" 
"--infer_data=/data/PaddlePaddle/1G/imagenet/val.bin" 
"--warmup_batch_size=100" "--batch_size=50" 
"--paddle_num_threads=28" 
"--iterations=1000"
`

> Besides, I run with `OMP_NUM_THREADS=1`.
I will try with this config.