@bingyanghuang 2019-08-01
I did the test in our CLX6248 with commit 233746d89d38486911ba28255c5bad7cd32f32f4 , and command line is: 
```shell
./paddle/fluid/inference/tests/api/test_analyzer_int8_image_classification --infer_model=third_party/inference_demo/int8v2/googlenet/model --infer_data=/~/.cache/paddle/dataset/int8/download/int8_full_val.bin --batch_size=1 --paddle_num_threads=1
```
Got the following results.
- *Before*
```
I0801 08:46:15.091809 262332 tester_helper.h:462] --- Performance summary ---
I0801 06:41:59.026803 249389 tester_helper.h:463] FP32: avg fps: 43.5602, avg latency: 22.9567 ms
I0801 06:41:59.026811 249389 tester_helper.h:466] INT8: avg fps: 93.9150, avg latency: 10.6479 ms
I0801 06:41:59.038902 249389 tester_helper.h:447] --- Accuracy summary ---
I0801 06:41:59.038915 249389 tester_helper.h:448] Accepted top1 accuracy drop threshold: 0.01. (condition: (FP32_top1_acc - INT8_top1_acc) <= threshold)
I0801 06:41:59.038921 249389 tester_helper.h:451] FP32: avg top1 accuracy: 0.7050
I0801 06:41:59.038925 249389 tester_helper.h:453] INT8: avg top1 accuracy: 0.7008
```
- *After*
```
I0801 08:46:15.091809 262332 tester_helper.h:462] --- Performance summary ---
I0801 08:46:15.091835 262332 tester_helper.h:463] FP32: avg fps: 44.4318, avg latency: 22.5064 ms
I0801 08:46:15.091846 262332 tester_helper.h:466] INT8: avg fps: 117.5707, avg latency: 8.5055 ms
I0801 08:46:15.101172 262332 tester_helper.h:447] --- Accuracy summary ---
I0801 08:46:15.101187 262332 tester_helper.h:448] Accepted top1 accuracy drop threshold: 0.01. (condition: (FP32_top1_acc - INT8_top1_acc) <= threshold)
I0801 08:46:15.101194 262332 tester_helper.h:451] FP32: avg top1 accuracy: 0.7050
I0801 08:46:15.101199 262332 tester_helper.h:453] INT8: avg top1 accuracy: 0.7003
```
Same conclusion as luotao. @wozna wojtek is planning to investigate the server configuration problem, will figure out the reason why we got the different results.