@bingyanghuang 2019-08-13
Based on luotao's benchmark on 6271:
![image](https://user-images.githubusercontent.com/33643817/62922047-06bd4480-bddd-11e9-828d-ec31b12b4a49.png)

INT8 performance got about 21% speedup but INT8 accuracy got 0.0005 accuracy drop . Since 0.0005 is minor compared  with big performance gain, we decided to merge this PR @Sand3r- @wojtuss @luotao1